requirements.txt
bs4==0.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
environs==9.5.0
flake8==6.1.0
openpyxl==3.1.2
pandas==2.1.0
pip-chill==1.0.3
pysocks==1.7.1
redis==5.0.0
requests==2.31.0
selenium==4.12.0
typing-extensions==4.7.1